Move responsibility for setting up hardware to HWAccelContext.
We will not likely continue down this path, as `man ffmpeg` makes the
point that CPU decoding is about the same speed as GPU decoding, and so
is only really of use for playback. I don't think PyAVs target is such
high performance playback, so we don't need to make the design
concesions required for this branch.

NOTE: This has not been tested to work. Two commits back is the original
PR squashed into a single commit and is more likely to work, although if
there is any hope of this being merged it will have to look more like
this commit does.

See (and further any discussion) #331 on GitHub.